requirements.txt
Flask
gunicorn
python-decouple
Flask-SQLAlchemy
pymysql
cryptography